weight-backend/requirements-dev.txt
Flask
Flask-CORS
Flask-SQLAlchemy
psycopg2-binary
pytest
